Refactor the updateInputs method in State.js
Specifically, I extracted the inner update() function out to the class
level (i.e. it is now on the State prototype). This made more sense to
me than passing in a 'State' variable to a function inside the state
class!

I also have a severe distaste for switch statements, so I refactored the
switch statement into an object of functions with the keys as the
property that was previously being 'switch'ed on.