Fix a bug in Pan.js causing exceptions
The move() function of the Pan recognizer was returning non-null values
at times when it should have returned null. Specifically, it was
returning an object containing an empty 'data' array. This was causing
problems, as the ZingTouch API currently provides a guarantee that
'data' will _not_ be empty when the handler is called! Therefore my use
of the API assumed the data would be there.

This fix ensures that the guarantee is upheld, by returning null if the
data array is empty, instead of returning the empty array.