Use a WeakSet to track initial event elements
We're checking to see wether an event initially occurred inside an
element. This was done by checking whether the initial event coordinates
are inside the element, as the element is currently situated. This is
dangerous, because the element could have been moved during that time.

As an alternative approach, I'm storing all the elements along the
event's propagation path inside a WeakSet. With this technique, we can
check whether an event was fired inside an element by checking if the
element is contained in the WeakSet.

The inspiration behind using a WeakSet is to allow fast lookups, and to
allow garbage collection on the elements to occur if they are removed
from the page.