Fix input tracking to allow resumption of inputs
Previously, a number of situations were causing the inputs array to be
emptied. This prevented a user from doing this like placing two fingers
on the device to interact, lifting one finger, then placing another
finger on the device.

This is a pretty hefty commit, but it fixes this situation.

There are likely to be bugs remaining in this implementation, and they
will need to be sought out and fixed.