Turn on smoothing by default in Pan
Given that the testing I did when I thought it was disabled turns out to
have had it enabled, it appears that I prefer having it enabled.

There remains a catch to this of course, namely that smoothing produces
awkward artifacts when used with a precise pointer device such as a
mouse. Will need to figure out some way of balancing those concerns.